% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.r
\docType{class}
\name{chickendipR-class}
\alias{chickendipR-class}
\title{Class "chickendipR" of chickendipR recipes}
\description{
A chickendipR recipe is represented as a S4 object
}
\section{Objects from the Class}{
 Objects are created by calls to
\code{\link{default_recipe}}
}

\examples{

showClass("chickendipR")
methods(class="chickendipR")
}
\keyword{classes}
