% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print}
\alias{print}
\alias{print.chickendipR}
\title{Print out the recipe with directions and notes}
\usage{
\method{print}{chickendipR}(x)
}
\description{
Print chickendipR recipe
}
\examples{
rec <- default_recipe()
print(rec)
}
