% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.r
\name{recipe}
\alias{recipe}
\title{Get default chickendipR recipe}
\usage{
recipe(scale = 2)
}
\arguments{
\item{scale:}{numeric to scale recipe by}
}
\description{
Instantiates the default chickendipR recipe
}
\examples{
rec <- recipe()
print(rec)
}
